Major Project/run.py
"""
Entry point for the Smart Attendance System.
Run:  python run.py
"""
from dotenv import load_dotenv
load_dotenv()

from app import create_app

app = create_app()

if __name__ == "__main__":
    app.run(host="127.0.0.1", port=5000, debug=True)
