Major Project/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.2
Werkzeug==3.1.3
SQLAlchemy==2.0.36

# Face Recognition & Computer Vision
# Windows: dlib is installed separately via pre-built wheel (see setup_windows.bat)
# macOS/Linux: pip install cmake && pip install dlib
face-recognition
opencv-python
numpy==1.26.4
scipy>=1.11.0,<1.15
Pillow>=10.0.0

# Export & Data
openpyxl==3.1.5
pandas==2.2.3

# Utilities
python-dotenv==1.0.1

# Required for face_recognition_models on Python 3.12+
setuptools<71